refactor: Add a new function to verify user input
A new function called "check_data_entry" is added to the codes. This
code verifies if a user's input is among a list of acceptable input
and prompt the user to retype if the original input is not valid.

This will help reducing redundancy in the "get filter" function and
other parts of the original code where user input needs to be verified.
Note that further changes in other parts of the code need to be made
to incorporate this new function.